some changes in files + some work on Put methods in react
