+++ ClientApp/src/App.js
@@ -6,7 +6,7 @@ import Counter from "./components/Counter";
 import FetchData from "./components/FetchData";
 import LoginPage from "./components/auth/login/LoginPage";
 import UsersListFetchData from "./components/UsersListFetchData";
-import TagList from "./components/tag/TagList";
+import TagsPage from "./components/tag/TagsPage";
 
 export default () => (
   <Layout>
@@ -14,7 +14,7 @@ export default () => (
     <Route path="/counter" component={Counter} />
     <Route path="/fetchdata/:startDateIndex?" component={FetchData} />
     <Route path="/login" component={LoginPage} />
-    <Route path="/tags" component={TagList} />
+    <Route path="/tags" component={TagsPage} />
     <Route path="/users/:startDateIndex?" component={UsersListFetchData} />
